requirements.txt
mysql-connector-python
openpyxl
pandas
requests
python-dotenv
